Continuing table source reorganization. Fixed typo in error message.
