+++ src/main/scala/org/openhorizon/exchangeapi/route/administration/Status.scala
@@ -15,9 +15,9 @@ import org.openhorizon.exchangeapi.table.node.agreement.NodeAgreementsTQ
 import org.openhorizon.exchangeapi.table.node.message.NodeMsgsTQ
 import org.openhorizon.exchangeapi.table.user.UsersTQ
 import org.openhorizon.exchangeapi.{Access, AuthenticationSupport, ExchMsg, HttpCode, TAction}
-import org.openhorizon.exchangeapi.table.SchemaTQ
 import org.openhorizon.exchangeapi.table.agreementbot.agreement.AgbotAgreementsTQ
 import org.openhorizon.exchangeapi.table.agreementbot.message.AgbotMsgsTQ
+import org.openhorizon.exchangeapi.table.schema.SchemaTQ
 import slick.jdbc.PostgresProfile.api._
 
 import scala.concurrent.ExecutionContext
