+++ src/main/scala/org/openhorizon/exchangeapi/route/managementpolicy/GetManagementPoliciesResponse.scala
@@ -1,6 +1,6 @@
 package org.openhorizon.exchangeapi.route.managementpolicy
 
-import org.openhorizon.exchangeapi.table.ManagementPolicy
+import org.openhorizon.exchangeapi.table.managementpolicy.ManagementPolicy
 
 final case class GetManagementPoliciesResponse(managementPolicy: Map[String, ManagementPolicy],
                                                lastIndex: Int)
