+++ src/main/scala/org/openhorizon/exchangeapi/route/managementpolicy/ManagementPoliciesRoutes.scala
@@ -15,6 +15,7 @@ import jakarta.ws.rs.{DELETE, GET, POST, PUT, Path}
 import org.json4s.jackson.Serialization.write
 import org.json4s.{DefaultFormats, Formats}
 import org.openhorizon.exchangeapi.auth.{AccessDeniedException, DBProcessingError}
+import org.openhorizon.exchangeapi.table.managementpolicy.{ManagementPoliciesTQ, ManagementPolicy}
 import org.openhorizon.exchangeapi.table.{organization, resourcechange, _}
 import org.openhorizon.exchangeapi.table.resourcechange.{ResChangeCategory, ResChangeOperation, ResChangeResource, ResourceChange}
 import org.openhorizon.exchangeapi.{Access, ApiRespType, ApiResponse, ApiTime, AuthCache, AuthenticationSupport, ExchConfig, ExchMsg, ExchangePosgtresErrorHandling, HttpCode, IUser, OrgAndId, TManagementPolicy}
