+++ src/main/scala/org/openhorizon/exchangeapi/table/deploymentpattern/OneSecretBindingService.scala
@@ -1,0 +1,7 @@
+package org.openhorizon.exchangeapi.table.deploymentpattern
+
+final case class OneSecretBindingService(serviceOrgid: String,
+                                         serviceUrl: String,
+                                         serviceArch: Option[String],
+                                         serviceVersionRange: Option[String],
+                                         secrets: List[Map[String, String]])
