+++ src/main/scala/org/openhorizon/exchangeapi/table/deploymentpattern/OneUserInputService.scala
@@ -1,0 +1,7 @@
+package org.openhorizon.exchangeapi.table.deploymentpattern
+
+final case class OneUserInputService(serviceOrgid: String,
+                                     serviceUrl: String,
+                                     serviceArch: Option[String],
+                                     serviceVersionRange: Option[String],
+                                     inputs: List[OneUserInputValue])
