+++ src/main/scala/org/openhorizon/exchangeapi/table/deploymentpattern/OneUserInputValue.scala
@@ -1,0 +1,4 @@
+package org.openhorizon.exchangeapi.table.deploymentpattern
+
+final case class OneUserInputValue(name: String,
+                                   value: Any)
