+++ src/main/scala/org/openhorizon/exchangeapi/table/deploymentpattern/PServices.scala
@@ -1,0 +1,9 @@
+package org.openhorizon.exchangeapi.table.deploymentpattern
+
+final case class PServices(serviceUrl: String,
+                           serviceOrgid: String,
+                           serviceArch: String,
+                           agreementLess: Option[Boolean],
+                           serviceVersions: List[PServiceVersions],
+                           dataVerification: Option[Map[String,Any]],
+                           nodeHealth: Option[Map[String,Int]])
