+++ src/main/scala/org/openhorizon/exchangeapi/table/deploymentpattern/key/PatternKey.scala
@@ -1,0 +1,4 @@
+package org.openhorizon.exchangeapi.table.deploymentpattern.key
+
+final case class PatternKey(key: String,
+                            lastUpdated: String)
