+++ src/main/scala/org/openhorizon/exchangeapi/table/deploymentpolicy/BService.scala
@@ -1,0 +1,8 @@
+package org.openhorizon.exchangeapi.table.deploymentpolicy
+
+final case class BService(name: String,
+                          org: String,
+                          arch: String,
+                          serviceVersions: List[BServiceVersions],
+                          nodeHealth: Option[Map[String,Int]],
+                          clusterNamespace: Option[String] = None)
