+++ src/main/scala/org/openhorizon/exchangeapi/table/deploymentpolicy/BServiceVersions.scala
@@ -1,0 +1,5 @@
+package org.openhorizon.exchangeapi.table.deploymentpolicy
+
+final case class BServiceVersions(version: String,
+                                  priority: Option[Map[String,Int]],
+                                  upgradePolicy: Option[Map[String,String]])
