+++ src/main/scala/org/openhorizon/exchangeapi/table/deploymentpolicy/search/SearchOffsetPolicyAttributes.scala
@@ -1,0 +1,6 @@
+package org.openhorizon.exchangeapi.table.deploymentpolicy.search
+
+final case class SearchOffsetPolicyAttributes(agbot: String,
+                                              offset: Option[String] = None,
+                                              policy: String,
+                                              session: Option[String] = None)
