+++ src/main/scala/org/openhorizon/exchangeapi/table/managementpolicy/AgentUpgradePolicy.scala
@@ -1,0 +1,6 @@
+package org.openhorizon.exchangeapi.table.managementpolicy
+
+final case class AgentUpgradePolicy(allowDowngrade: Boolean,
+                                    manifest: String) {
+  override def clone = new AgentUpgradePolicy(allowDowngrade, manifest)
+}
