+++ src/main/scala/org/openhorizon/exchangeapi/table/schema/Schema.scala
@@ -1,0 +1,6 @@
+package org.openhorizon.exchangeapi.table.schema
+
+final case class Schema(id: Int,
+                        schemaVersion: Int,
+                        description: String,
+                        lastUpdated: String)
