+++ src/test/scala/org/openhorizon/exchangeapi/route/agentconfigurationmanagement/TestDeleteAgentConfigMgmt.scala
@@ -10,7 +10,7 @@ import org.openhorizon.exchangeapi.route.organization.{ChangeEntry, PatchOrgRequ
 import org.openhorizon.exchangeapi.route.user.PostPutUsersRequest
 import org.openhorizon.exchangeapi.table.agreementbot.message.{AgbotMsgRow, AgbotMsgsTQ}
 import org.openhorizon.exchangeapi.table.agreementbot.{AgbotRow, AgbotsTQ}
-import org.openhorizon.exchangeapi.table.deploymentpolicy.{SearchOffsetPolicyAttributes, SearchOffsetPolicyTQ}
+import org.openhorizon.exchangeapi.table.deploymentpolicy.search.{SearchOffsetPolicyAttributes, SearchOffsetPolicyTQ}
 import org.openhorizon.exchangeapi.table.node.{NodeRow, NodesTQ}
 import org.openhorizon.exchangeapi.table.organization.{OrgRow, OrgsTQ}
 import org.openhorizon.exchangeapi.table.resourcechange.{ResChangeCategory, ResChangeOperation, ResChangeResource, ResourceChangeRow, ResourceChangesTQ}
