+++ src/test/scala/org/openhorizon/exchangeapi/route/managementpolicy/ManagementPoliciesSuite.scala
@@ -6,10 +6,10 @@ import org.json4s.{DefaultFormats, jvalue2extractable}
 import org.junit.runner.RunWith
 import org.openhorizon.exchangeapi.route.organization.{ChangeEntry, ResourceChangesRequest, ResourceChangesRespObject}
 import org.openhorizon.exchangeapi.route.user.PostPutUsersRequest
+import org.openhorizon.exchangeapi.table.managementpolicy.ManagementPolicy
 import org.openhorizon.exchangeapi.table.node.{NodeRow, NodesTQ}
 import org.openhorizon.exchangeapi.table.organization.{OrgRow, OrgsTQ}
 import org.openhorizon.exchangeapi.table.service.OneProperty
-import org.openhorizon.exchangeapi.table.ManagementPolicy
 import org.openhorizon.exchangeapi.table.resourcechange.ResourceChangesTQ
 import org.openhorizon.exchangeapi.{ApiResponse, ApiTime, ApiUtils, HttpCode, Role, TestDBConnection}
 import org.scalatest.BeforeAndAfterAll
